pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bandwagon"
version = "0.1.0"
description = "Bulk photometric catalog cross-matching helpers for building SED tables."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "astropy>=6",
    "astroquery>=0.4.8",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
]
spectra = [
    "sparclclient>=1.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/bandwagon"]
